Add links to archived copy of GATK 3.7 docs
In 2021 the team behind GATK restructured their website and stopped maintaining documentation for GATK3 (https://gatk.broadinstitute.org/hc/en-us/articles/360038154292-Welcome-to-the-new-home-of-GATK). The documentation now lives on Github (https://github.com/broadinstitute/gatk-docs/tree/master), but is unfortunately stored as raw HTML files and is thus not easily viewable in a browser.

The website https://raw.githack.com is a free tool that acts as a caching proxy that re-serves the HTML files with an appropriate `Content-Type`. I've used it to wrap the following two URLs:

https://github.com/broadinstitute/gatk-docs/blob/8fcf44bb0686f2f7d442aade181ff6ed508a97de/gatk3-tooldocs/3.7-0/org_broadinstitute_gatk_tools_walkers_indels_IndelRealigner.html
https://github.com/broadinstitute/gatk-docs/blob/8fcf44bb0686f2f7d442aade181ff6ed508a97de/gatk3-tooldocs/3.7-0/org_broadinstitute_gatk_tools_walkers_indels_RealignerTargetCreator.html